requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
animec
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
discord==1.7.3
discord.py==1.7.3
Flask==1.1.2
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lxml==4.7.1
MarkupSafe==2.0.1
multidict==6.0.2
parsel==1.6.0
prawcore==2.3.0
priority==2.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dotenv==0.19.2
python-telegram-bot==13.11
qrcode==7.3.1
queuelib==1.6.2
replit==1.4.3
requests==2.27.1
Scrapy==2.4.1
service-identity==21.1.0
six==1.16.0
sqlitedict==1.7.0
Twisted==21.7.0
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.8
viberbot==1.0.12
w3lib==1.22.0
websockets==3.4
Werkzeug==2.0.2
yarl==1.7.2
youtube-dl==2021.12.17
zipp==3.7.0
zope.interface==5.4.0
